server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Automat==22.10.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
distro==1.9.0
fastembed==0.2.5
filelock==3.13.1
Flask==3.0.2
flatbuffers==24.3.25
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
loguru==0.7.2
lxml==5.1.0
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==1.26.4
onnx==1.16.0
onnxruntime==1.17.1
openai==1.14.3
packaging==23.2
parsel==1.8.1
pinecone-client==3.2.2
Protego==0.3.0
protobuf==5.26.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyDispatcher==2.0.7
pyOpenSSL==24.0.0
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
scikit-learn==1.4.1.post1
scipy==1.12.0
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sympy==1.12
threadpoolctl==3.4.0
tldextract==5.1.1
tokenizers==0.15.2
tqdm==4.66.2
Twisted==24.3.0
typing_extensions==4.10.0
urllib3==2.2.1
w3lib==2.1.2
Werkzeug==3.0.1
zope.interface==6.2
